fix: Use `contourArea` instead of pixel counting
Previously, the perimeter of an object was determined based on a contour
fit, whereas the area was determined by pixel counting (using `stats`).
Since we are measuring the perimeter of one object (the contour) but
count pixels on another (the binary image mask), there are sometimes
discrepancies between perimeter and area that lead to an unphysical
shape index S < 3.54.
See also:
https://docs.opencv.org/4.x/d3/dc0/group__imgproc__shape.html#ga2c759ed9f497d4a618048a2f56dc97f1

Fix these implausible results by measuring the area of the contour
instead of resorting to pixel counting.